add keymap dialog to embeds
