+++ web/experimental/new_playground/styles.scss
@@ -433,12 +433,6 @@ a {
   margin-left: 12px;
 }
 
-// Keys dialog
-.mdc-dialog__surface {
-  // Make the dialog a minimum of 400px to best display the keyboard dialog
-  min-width: 512px !important;
-}
-
 .keyboard {
   display: inline-block;
   background: url('../../pictures/keyboard.svg') center no-repeat;
@@ -452,35 +446,6 @@ a {
   &:hover {
     opacity: 1;
   }
-}
-
-.keys-dialog {
-  top: auto !important;
-  bottom: 20px;
-
-  dl {
-    padding: 0.5em;
-  }
-
-  dt {
-    float: left;
-    clear: left;
-  }
-  dd {
-    text-align: right;
-  }
-  dd > span {
-    font-family: "Helvetica Neue", "Inconsolata", monospace;
-    padding: 0 4px;
-    font-size: 10pt;
-    color: $dark-blue;
-    border: 1px solid $dark-blue;
-    border-radius: 2px;
-    margin: 2px 2px;
-    white-space: nowrap;
-    display: inline-block;
-  }
-
 }
 
 // Unread console counter
